pyproject.toml
[project]
name = "poetry-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.7",
    "openai-agents>=0.2.4",
    "python-dotenv>=1.1.1",
]
